Move treadle dependency to the tests
Use inheritance to make TesterDriver Backend API extensible, then define
a TreadleBackend in the test project
