+++ src/test/scala/chisel3/testers/TestUtils.scala
@@ -1,0 +1,12 @@
+// See LICENSE for license details.
+
+package chisel3.testers
+
+import TesterDriver.Backend
+import firrtl.AnnotationSeq
+
+object TestUtils {
+  // Useful because TesterDriver.Backend is chisel3 package private
+  def containsBackend(annos: AnnotationSeq): Boolean =
+    annos.collectFirst { case b: Backend => b }.isDefined
+}
